Remove uppercasing of HTTP Status reasons
This code is triggered when creating a Response with an int as status
(`return Response("foo", 404)`).

-> HTTP/1.1 404 NOT FOUND

Calling upper() seems unnecessary and is inconsistent with the behavior
when you specify the reason by calling Response() with a string as 
status. (it's also unusual IMO)
-> HTTP/1.1 404 Not Found (or whatever was specified)

Originally from commit 85e43c12fd74375779d23c117594ea325b944d8c
https://github.com/mitsuhiko/werkzeug/commit/85e43c12fd74375779d23c117594ea325b944d8c