Changed .data and .get_data() again
This restores the old pre 0.9 behavior for the data attribute
on request objects and it gives the .get_data() method a
parameter to have that same behavior.  In addition to that a
new method was added that creates a stream for parsing that
honors previously cached data.  This is now also being used
by the form data parser.