Explain how to quit the debug server
Many new users start the server in debug mode locally, then don't know how to
kill it or "get their terminal back" after the Werkzeug server starts. This
adds a short message explaining how to quit the server and "get your terminal
back". CTRL+C works on Mac and Linux and according to the IRC channel, Windows
as well, otherwise we'd probably want to customize the message depending on the
host OS.

This message is similar to Python's "Use quit() or Ctrl-D (i.e. EOF) to exit"
when you type "quit" in the REPL.