Add failing test case to show broken Rule issue in 0.10
This was introduced in 426b98534fa623294100c4b14fe0d8fd7fa66e58 which
properly instantiates the subclass of a Rule instead of returning a Rule
object.  The issue is that empty() uses args instead of kwargs, so
anyone relying on kwargs to inject a new value will have bugs
introduced.