Fix this test case so it proves something, not simply the absence
Was thinking about this and it bugged me that this test didn't do
anything positive to ensure it passed.  A silently passing test with no
explicit failure test case will be overlooked later and possibly
removed.  This provides an explicit state that shows the failure along
with a sanity check for empty().