Make websockets a bit nicer.
To create a websocket it is necessary to set the connection and upgrade
headers and not set the content-length header. At the moment werkzeug
will set a second connection header for close in this situation, this as it
happens doesn't seem to bother the couple of browser I've tried this with
(Chrome and IE) but it's not pretty.