Routing - try next rule without redirect.
Try the next rule without redirect if the current request method is not match for the current rule methods.
For POST, PUT and PATCH methods redirect is not allowed.

Sample code:
```python
@app.route('/path/', methods=['GET'])
@app.route('/path',  methods=['GET'])
@auth.login_required
def get():
    return make_response(do_get())

@app.route('/path/', methods=['POST'])
@app.route('/path',  methods=['POST'])
@jsonschema.validate('path', 'create')
@auth.login_required
def create():
    return make_response(do_create(request))

@app.route('/path/', methods=['DELETE'])
@app.route('/path',  methods=['DELETE'])
@auth.login_required
def delete():
    return make_response(do_delete(request))

@app.route('/path/', methods=['PATCH'])
@app.route('/path',  methods=['PATCH'])
@auth.login_required
def edit():
    id = request.args.get('id')
    return make_response(do_edit(id, request))
```

Sample error without fix for PATCH method:
FormDataRoutingRedirect: A request was sent to this URL (http://localhost:9000/path?id=12345) but a redirect was issued automatically by the routing system to "http://localhost:9000/path/?id=12345".  The URL was defined with a trailing slash so Flask will automatically redirect to the URL with the trailing slash if it was accessed without one.  Make sure to directly send your PATCH-request to this URL since we can't make browsers or HTTP clients redirect with form data reliably or without user interaction.