fix: environ might be missing
how it happened?
the environ is attached to WSGIRequestHandler object in run_wsgi method. But run_wsgi is only executed after the request is pasted in handle_one_request method. If request is pasted failed and try to send_error message. It will try to call address_string method to get the client address and raise the AttributeError(AttributeError: 'WSGIRequestHandler' object has no attribute 'environ')