Create a thread to reap death process
ForkingWSGIServer use `SocketServer.ForkingMixIn` to implement a
multiprocess server.  This class provides a workflow that collect
death process (process in Zombie status) before the
`process_request`.  This means that this process itself will be
in Zombie status at the end of the request, that will be eventually
collected during the next `process_request`.

To minimize transient Zombie process, `ForkingWSGIServer` is
creating a daemon thread (via `threading.Timer`) to call the
collector every (by default) 5 seconds.

Fixes #810