Two more changes towards recognizing the correct useragents:
Recognize freebsd platform.
Move seamonkey in the list of browsers above firefox, because most of
the useragent strings for seamonkey contain firefox, which results in
the browser being recognized as firefox instead of seamonkey.
Remove 'mozilla' from the regex for seamonkey as it results in mozilla
version being recognized as the seamonkey version.