Add an extras_require entry for the optional watchdog dependency
The reloader feature has an optional dependency on watchdog:
http://werkzeug.pocoo.org/docs/0.11/serving/#reloader

This change allows the use of `Werkzeug[watchdog]` in requirements
files, which makes the reason behind the dependency clearer, and also
allows for version specifiers to be more easily introduced in the future,
should there be incompatibilities with certain versions of watchdog.

Fixes #929.