check file size (#959)
* check file size

In a frozen app, using 'pkg_resources' the file size returned is zero, and with development server the files/images are not shown. With cherrypy wsigserver, throw a exception with ValueError about content-length. With this patch, all is fine.
With this patch, exclusively when content-length is zero, this information isn't passed and firefox and chrome accept the image/scripts/etc and show all.