Update _legal_cookie_chars to include / and =
Both / and = may occur in a cookie value without requiring that the value be quoted.
Reference https://www.ietf.org/rfc/rfc6265.txt

While the presence of added quotes should be inconsequential, this is
motivated by interacting with a server that did not seem to unquote the
cookie value for an auth token, which strictly speaking did not need to
be quoted.