Catch StopIteration as a valid case. (#1280)
According to the WSGI spec[1], it is a valid case to both:

a) lazily call `start_response` in request processing

> this invocation may be performed by the iterable’s first iteration, so
> servers must not assume that start_response() has been called before
> they begin iterating over the iterable

b) return an empty iterator (e.g. in the case of a 204 No Content)

> When called by the server, the application object must return an
> iterable yielding zero or more strings.

When both of these were true, previously, Werkzeug's test client would
propagate a `StopIteration` exception from the WSGI app's returned
iterable. Since this is a valid case, we now catch the `StopIteration`
exception if `start_response` has _also_ been called. If
`start_response` has not been called we re-raise, as this is not a valid
case under the WSGI spec, and as mentioned in the docstring of this
function...

> If passed an invalid WSGI application the behavior of this function is
> undefined.

[1]: https://www.python.org/dev/peps/pep-0333/#specification-details

Co-Authored-By: Alistair Lynn <arplynn@gmail.com>