Unicodify the header name (key) as well as value
In Python 2 using a Headers object as so,

    h = Headers()
    h[b'X-Foo'] = b'something'

gives

    X-Foo: something

in the actual HTTP message, whereas in Python 3 it is

    b'X-Foo': something

this is because only the value is currently unicodified. This change
ensures the header name (key) is also unicodified as well therefore
matching the Python 2 usage and the expected usage.

This also removes a regression test for unicode header keys that is
not required. It was recognised in flask#758 as resulting in a HTTP
message containing,

    u'X-Foo': something

however this behaviour was fixed in
f3435a346d0d559d1c19bbfca3745476c4942124 introducing the test, broken
in 6049a4ff3c5224bedd24cdb29a23a5f1a9c2bf43 and then comprehensively
fixed in db00dfb41be37ff34e4014ee4a31d9abac8165e2. Therefore with the
last commit and tests as added the regression test is not required.