Conserve order for equal quality in accept datastructures
Both RFC4647-2.3 and RFC7231 imply that user agents may put equal
quality tags (especially language tags) in order of their prefered
priority. This means that this would ideally be the case,

    a = parse_accept_header("en-US,fr-FR", LanguageAccept)
    assert a.best == "en-US"

which it previously wasn't as `f` is after `e`.

I've looked back through the commits and found the previous behaviour
(sorting by tag name) was added with Accept in
802a12c0850439088de46ae188c64ab259b16695. I can find no reason for the
previous behaviour to be requirement.