Improve the typing based on checking against Quart
The URLPartsTuple limits the allowed type accepted by the functions to
a 5 part tuple of either all bytes or all str.

Correct expires such that it also accepts a float timestamp and
max_age such that is also accepts a timedelta.

Use overloads to make it explicit what the return type is based on the
value of the cls argument.

The additional ignore in debug is related to this and due to,

    src/werkzeug/debug/__init__.py:385: error: No overload variant of "parse_cookie" matches argument type "Dict[str, Any]"

but it is not clear why.

The methods is acceptable for it to be an iterable of str, note that
an error is raised if a str is passed.