test Client always returns a TestResponse
TestResponse.request is a Request for the environ that created the
response. TestResponse.history is a list of intermediate responses
when following redirects.

Client.response_wrapper is always a subclass of TestResponse. If no
wrapper is passed or it's Response or BaseResponse, use TestResponse.
Otherwise, create a subclass with the wrapper and TestResponse.

EnvironBuilder.get_request returns a Request instead of BaseRequest.

Deprecate Client.open(as_tuple=True). Treating the response as a tuple
(unpacking, indexing) raises a deprecation warning.

Update all tests to use TestResponse instead of tuple. Fix the ProxyFix
test that was relying on mutating the environ. Update the
follow_redirects history test to follow multiple redirects.