Split the Sans-IO parts out of the Request wrapper
This might be clearer if you consider the names as `BaseRequest` and
`WSGIRequest` (although these names aren't possible for compatibility
reasons). The `SansIORequest` contains all the information that is not
WSGI or IO dependent thereby allowing it to be used in ASGI projects,
notably Quart.

This has a very slight performance impact as the path, scheme,
headers, and method are no longer lazily loaded. However, this impact
is likely just a dictionary lookup.

The changes in other files are minor typing improvements related to
the typing changed in the request wrappers.