Refactor the Multipart parsing into a Sans-IO layer
This allows it to be used in async (ASGI) frameworks. It also
hopefully makes the code a little clearer to follow.

This removes the ``Content-Transfer-Encoding`` support as RFC7578
states it is deprecated and

    Currently, no deployed implementations that send such bodies have
    been discovered.

This removes the IE6 name fix as IE6 is no longer supported (either
directly or by Werkzeug).

This requires the dataclasses backport, which I think is ok as
dataclasses are in the stdlib, and 3.6 has less than a year till EoL.

The API is based on that successfully used by h11.