Fix LocalStack mutable list issue
The stack stored by a LocalStack should be inherited from any parent
context, but crucially and changes to it should be local to its
context. This didn't happen previously as lists are mutable and the
list instance was being shared between the contexts. This change, much
like the local dictionary usage, now takes a copy and more explicitly
separates the contexts.

Thanks to @miguelgrinberg for pointing out this issue.