Bugfix the calculation of weighting based on static parts
Only static parts that are more than a `/` should count towards the
weighting, as they influence the part for matching. This prevents
`/<path:p>/<a>/<b>` matching before `/<path:p>/b` as only the latter
rule should have a static weighting as it is the only rule that has a
static part.