Fix strict slash redirects for final RuleParts
Final RuleParts, for example ones with a path converter will consume
the remaining request target trying to match. This would raise
NotFound if the rule required a trailing slash and the request target
was missing one. Now, with this fix, it will redirect instead.

This is implemented via a suffix group and marker to final
RuleParts. The suffix group will optionally allow for the trailing
slash to be present, and if missing (and required) will trigger a
redirect response.

Note that rules like `/<path:path>` and `/<path:path>/` are not
distinguishable as the former matches the latter. Hence the
strict_slashes False behavior for other rules does not apply.