Allow the router to cope with nested groups
The previous, regex table, version of the router wrapped each
converter's regex in a named capturing group which was then extracted
from the match and passed to the converter as the value with the name
matching the rule variable name. This meant that any nested groups
within that named group were ignored.

The new, state machine, version of the router couldn't use this
approach as the variable names are likely unique and hence the same
rule with different variable names would have seperate states, rather
than the same. Therefore the state machine version simply took the
groups in order and passed them to the converters in order.

The requirement that the groups be in order means that nested groups
cannot be used, as the nested groups are not ignored and in fact
result in more group matches than there are converters.

As Werkzeug/Flask did support nested groups (although not directly as
they aren't used in Werkzeug/Flask) it makes sense to restore this
support. To do so I've used the previous approach with a generic name
`__werkzeug_#` that is unlikely to conflict with any user defined
name. Then only matching only the named groups (with the
`__werkzeug_#` prefix) are passed to the converters in the order given
by the name i.e. `__werkzeug_0` for the first converter in the part,
`__werkzeug_1` for the second etc... This numbering is required as
the match groupdict() order is not guaranteed anywhere.