Fix the parsing of large multipart bodies
There were two issues to fix. Firstly if a boundary couldn't be found
the parser should have parsed up to the end of the buffer or last
newline (whichever is earlier). However the last newline would be the
first character since 082e0e5b9c01fa3178ac0153413f082616f10914 as the
DATA_START state would have a buffer that starts with newline. This
was fixed by changing the last_newline method to take the data to
search as an argument.

Secondly the parsing was slow as the shortcut search for the boundary
was removed resulting in full regex matches on each
iteration. Restoring the shortcut restores the previous performance.